model/requirements.txt
adal==1.2.7
aiofiles==22.1.0
aiohttp==3.8.3
aiokafka==0.8.0
aiosignal==1.3.1
alembic==1.9.2
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.2
azure-graphrbac==0.61.1
azure-mgmt-authorization==3.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.2.1
azure-mgmt-storage==20.1.0
azureml-core==1.48.0
backoff==2.2.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.1
bentoml==1.0.13
boto3==1.26.59
botocore==1.29.59
Brotli==1.0.9
build==0.10.0
cachetools==5.3.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
circus==0.18.0
click==8.1.3
click-option-group==0.5.5
cloudpickle==2.2.1
contextlib2==21.6.0
contourpy==1.0.7
cryptography==38.0.4
cycler==0.11.0
databricks-cli==0.17.4
deepmerge==1.1.0
Deprecated==1.2.13
distlib==0.3.6
docker==6.0.1
entrypoints==0.4
exceptiongroup==1.1.0
fastapi==0.88.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
gevent==22.10.2
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.1
grpcio==1.51.1
grpcio-health-checking==1.48.2
gunicorn==20.1.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
isodate==0.6.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
kafka-python==2.0.2
kiwisolver==1.4.4
knack==0.10.1
kubernetes==25.3.0
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
mlflow==2.1.1
mlserver==1.3.0.dev2
mlserver-mlflow==1.3.0.dev2
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
ndg-httpsclient==0.5.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opentelemetry-api==1.14.0
opentelemetry-exporter-otlp-proto-http==1.14.0
opentelemetry-instrumentation==0.35b0
opentelemetry-instrumentation-aiohttp-client==0.35b0
opentelemetry-instrumentation-asgi==0.35b0
opentelemetry-instrumentation-grpc==0.35b0
opentelemetry-proto==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-semantic-conventions==0.35b0
opentelemetry-util-http==0.35b0
orjson==3.8.5
packaging==21.3
pandas==1.5.3
paramiko==2.12.0
pathspec==0.11.0
Pillow==9.4.0
pip-requirements-parser==32.0.1
pip-tools==6.12.2
pkginfo==1.9.6
platformdirs==2.6.2
portalocker==2.7.0
prometheus-client==0.16.0
prometheus-flask-exporter==0.21.0
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
py-grpc-prometheus==0.7.0
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
PyNaCl==1.5.0
pynvml==11.4.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
python-multipart==0.0.5
python-rapidjson==1.9
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
querystring-parser==1.2.4
requests==2.28.2
requests-auth-aws-sigv4==0.7
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
schema==0.7.5
scikit-learn==1.2.1
scipy==1.5.2
SecretStorage==3.3.3
shap==0.41.0
simple-di==0.1.5
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
starlette==0.22.0
starlette-exporter==0.15.1
tabulate==0.9.0
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.4.0
tornado==6.2
tqdm==4.48.2
tritonclient==2.29.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
watchfiles==0.18.1
websocket-client==1.5.0
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
